#4서버작업
컨트롤러 없애고 서비스 기능을 util로 바꿈
